Use list of available network interfaces as default parameters
