+++ src/raspberrypi/devices/device_factory.h
@@ -33,12 +33,13 @@ public:
 
 	static DeviceFactory& instance();
 
+	Device *CreateDevice(PbDeviceType type, const string& filename, const string& ext);
+
 	const set<uint32_t>& GetSectorSizes(PbDeviceType type) { return sector_sizes[type]; }
 	const set<uint32_t>& GetSectorSizes(const string&);
 	const set<uint64_t> GetCapacities(PbDeviceType);
 	const map<string, string>& GetDefaultParams(PbDeviceType type) { return default_params[type]; }
-
-	Device *CreateDevice(PbDeviceType type, const string& filename, const string& ext);
+	const list<string> GetNetworkInterfaces() const;
 
 private:
