+++ src/raspberrypi/rascsi.cpp
@@ -32,7 +32,6 @@
 #include "spdlog/sinks/stdout_color_sinks.h"
 #include <spdlog/async.h>
 #include <sys/sendfile.h>
-#include <ifaddrs.h>
 #include <string>
 #include <sstream>
 #include <iostream>
@@ -558,36 +557,9 @@ void GetAvailableImages(PbImageFilesInfo& image_files_info)
 
 void GetNetworkInterfacesInfo(PbNetworkInterfacesInfo& network_interfaces_info)
 {
-	struct ifaddrs *addrs;
-	getifaddrs(&addrs);
-	struct ifaddrs *tmp = addrs;
-
-	while (tmp) {
-	    if (tmp->ifa_addr && tmp->ifa_addr->sa_family == AF_PACKET &&
-	    		strcmp(tmp->ifa_name, "lo") && strcmp(tmp->ifa_name, "rascsi_bridge")) {
-	        int fd = socket(PF_INET, SOCK_DGRAM, IPPROTO_IP);
-
-	        struct ifreq ifr;
-	        memset(&ifr, 0, sizeof(ifr));
-
-	        strcpy(ifr.ifr_name, tmp->ifa_name);
-	        if (!ioctl(fd, SIOCGIFFLAGS, &ifr)) {
-	        	close(fd);
-
-	        	// Only list interfaces that are up
-	        	if (ifr.ifr_flags & IFF_UP) {
-	        		network_interfaces_info.add_name(tmp->ifa_name);
-	        	}
-	        }
-	        else {
-	        	close(fd);
-	        }
-	    }
-
-	    tmp = tmp->ifa_next;
+	for (const auto& network_interface : device_factory.GetNetworkInterfaces()) {
+		network_interfaces_info.add_name(network_interface);
 	}
-
-	freeifaddrs(addrs);
 }
 
 void GetDevice(const Device *device, PbDevice *pb_device)
